Rename base.rs to builder.rs, move things related to the final CB to mod.rs
